Remove qajax, use axios consistently
So far, replaced Qajax, removed the dependency, checked that tests
aren't failing, and checked the interface superficially.  Next step is
to replace Q.Promise.

Looked into using fetch but a custom timeout and conversion of the
response to json is required, both of which axios does easily.  We could
implement these on top of fetch, but it seems cleaner this way.

The pull request that this commit will be part of works on #19 and #51

NOTE: axios is promising some good improvements in 0.19 but it's been
stuck in beta for a few months with no activity moving towards release.