Refactor uses of q to native Promise
This is a big refactor, so I'm pushing the first thing that builds, this
way people can see what's affected and let me know how to proceed.  My
next steps would be to test thoroughly, and I'm open to other ideas.
Let's work together to land this soon before it trips up on other
changes.

Ongoing work on #19 and #51.