fix table split time dimension limit
the current implementation is looking up the dimension for the split
using the first split in the list of splits, which is obviously
incorrect. this results in a bug when applying limits for time
dimensions, as unless the first split dimension kind is "time" it'll
change the limit to either 5 or 50.