Only render portal children after DOM element is inserted
As documented here https://reactjs.org/docs/portals.html, the portal element is
inserted in the DOM tree after the <BodyPortal>'s children are mounted, meaning
that they will be mounted on a detached DOM node. For children that manage focus
(like the <ClearableInput>) we must ensure they're only rendered after their
host element is inserted in the DOM tree.