Use true when setting 'trust proxy' on express (#668)
'trust proxy' was set to "1", which trusted only the right-most entry in
X-Forwarded-* headers.  This changes it to "true" so that the left-most
entry is used by express as the client ip.  So if XFF looks like this:

123.234.1.1, 10.1.1.10, 234.123.0.1

We would now see 123.234.1.1 as `req.ip`, which is probably more common
and expected.  Prior to this change it would have been 234.123.0.1

'trust proxy' guide here: expressjs.com/en/guide/behind-proxies.html
And the relevant call stack is:
  turnilo -> express (lib/request.js) -> proxy-addr -> forwarded