fix: consider user selected timezone while exporting the reports
problem:
Time in CSV exported reports are always in UTC timezone irrespective of user
selected timezone in the UI. There are use case when time in the exported
reports should be other timezone than UTC.

fix:
plywood passes the user selected timezone to format function here
https://github.com/implydata/plywood/blob/master/src/datatypes/dataset.ts#L343
We can use that to get time in the user selected timezone. The time in reports
will also have information about timezone because of ISO format.